requirements.txt
tensorflow-cpu==2.18.0
Flask==3.0.2
numpy==1.26.4
h5py==3.11.0
librosa==0.10.1
soundfile==0.12.1
scipy==1.13.0
serverless-wsgi==3.0.2
joblib==1.3.2
numba==0.58.1
llvmlite==0.41.1